backend/requirements.txt
# FastAPI和核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis缓存
redis==5.0.1
aioredis==2.0.1

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
typer==0.9.0
rich==13.7.0

# JSON处理
orjson==3.9.10

# 测试相关
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 类型提示
types-redis==4.6.0.11